+++ emulator-egui/src/main.rs
@@ -267,7 +267,8 @@ impl MyEguiApp {
                             for offset in 0..memory_view_info_lock.buffer.len() {
                                 memory_view_info_lock.buffer[offset] = cpu.bus.read_byte_address(
                                     memory_view_info_lock.offset + (offset as u32),
-                                );
+                                )
+                                    as u8;
                             }
                         }
